Issue fsync less frequently
We seem to be issueing fsync too frequently for write jobs. For a block size
of 4K, it will after writing 4MB of data, we issue a fsync.

Reduce fsync frequency and issue one after every 32K number of I/O. For
a block size of 4K, this will mean after every 128MB of data written.
  
Signed-off-by: Vivek Goyal <vgoyal@redhat.com>