Remove "fysnc" option from write jobs
This forces writing of all data on disk on host. This should not be a
must when we are trying to primarily measure the spped of operation
through virtiofs (and not how efficint host filesystem is). We can
very well cache all writes on host, and thats fine.

So remove it. If you need it, add it in your job file.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>