parse: Do not filter out results with 0 bandwidth
If a test/job has not been run, it returns "Unknown" as output and we filter
it out. 

But if a test has been run but it errored out because configuration does not
support it then 0 is returned. Don't filter it out. I think user should know
that this job was run but it failed to run. (ex. mmap jobs with cache=none).

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>