Revert back to 4G file size
I thought that if we have already read file size data (because of fast speed),
then we will re-read/re-write same location because of runtime restriction.
But that does not seem to be the case. Test will finish before runtime if
file size equivalent I/O is done before runtime elapses.

If that's the case, fallback to previous 4G per file size limit. It needs
less space for testing. Also dax window can be smaller (16G).

Signed-of-by: Vivek Goyal <vgoyal@redhat.com>